@pentschev 2019-12-05
rerun tests