@pentschev 2019-12-05
This error has been fixed upstream, could you merge branch-0.11 in your branch @rjzamora or would you like me to do it? Is this ready for review as well?