@quasiben 2019-12-05
@pentschev i think this is ready for review