@pentschev 2019-12-05
LGTM, I can't really comment much about the cuDF problem, but I trust both of you to be good with it. :)

Would you say this is ready to merge or are you pushing some commits still?