@rjzamora 2019-12-05
> Would you say this is ready to merge or are you pushing some commits still?

I'm not planning any changes at the moment, but I'm still open to suggestions - cc @mrocklin