@pentschev 2019-12-05
Ok, we can still get things merged until midnight, so I'll merge this in about an hour if there are no further comments, otherwise it will only be merged for 0.12.