@pentschev 2019-12-05
Actually, we need to update the https://github.com/rapidsai/dask-cuda/blob/branch-0.12/CHANGELOG.rst, I can do that if you don't mind me pushing to your branch @rjzamora, or if you prefer you can do it, just let me know.