@mrocklin 2019-12-05
> I'm not planning any changes at the moment, but I'm still open to suggestions - cc @mrocklin

Nothing from me